ALT-Scann8 1.9.29 - Move classes DynamicSpinbox and CameraResolutions to dedicated files
